Uploaded file manager protocol
